DE/crawling/requirements.txt
# module
click
python-dotenv
tqdm

# youtube api
google-api-python-client
oauth2client

# transcript api
youtube_transcript_api

# dataframe
pandas